dynamic-not-editable-pricing/requirements.txt
python-dotenv
polar-sdk
upstash-redis
fastapi[standard]